Move simplex control into simplex.ts
