+++ src/index.ts
@@ -1,4 +1,3 @@
 export { Brent, BrentControlParam, fitBrent } from "./brent";
-export { SimplexControlParam } from "./control";
-export { Simplex, fitSimplex } from "./simplex";
+export { Simplex, SimplexControlParam, fitSimplex } from "./simplex";
 export { Result, TargetFn } from "./types";
